api/requirements.txt
flask==1.1.1
flask_graphql==2.0.0
graphene==2.1.8
python-dotenv==0.10.3
